+++ Craftsman/Builders/Features/QueryGetListBuilder.cs
@@ -32,7 +32,6 @@ public class QueryGetListBuilder
         var repoInterfaceProp = $"{entity.Name.LowercaseFirstLetter()}Repository";
         
         var dtoClassPath = ClassPathHelper.DtoClassPath(srcDirectory, "", entity.Plural, projectBaseName);
-        var wrapperClassPath = ClassPathHelper.WrappersClassPath(srcDirectory, "", projectBaseName);
         var entityServicesClassPath = ClassPathHelper.EntityServicesClassPath(srcDirectory, "", entity.Plural, projectBaseName);
         var exceptionsClassPath = ClassPathHelper.ExceptionsClassPath(srcDirectory, "", projectBaseName);
         var resourcesClassPath = ClassPathHelper.WebApiResourcesClassPath(srcDirectory, "", projectBaseName);
@@ -51,7 +50,6 @@ public class QueryGetListBuilder
 
 using {dtoClassPath.ClassNamespace};
 using {entityServicesClassPath.ClassNamespace};
-using {wrapperClassPath.ClassNamespace};
 using {exceptionsClassPath.ClassNamespace};
 using {resourcesClassPath.ClassNamespace};{permissionsUsing}
 using Mappings;